feat(masthead): masthead changes
