+++ app/page.tsx
@@ -5,62 +5,21 @@ import { motion } from "framer-motion"
 import { Canvas } from '@react-three/fiber';
 import { OrbitControls } from '@react-three/drei';
 import { Suspense, useEffect, useState } from 'react';
-import { Glasses } from '@/components/Glasses';
+import { Glasses } from '@/components/masthead/Glasses';
 import Footer from '@/components/Footer';
 import { posts, projects } from '@prisma/client';
 import { useQuery } from 'react-query';
 import Project from '@/components/Project';
 import BlogPreview from '@/components/BlogPreview';
-import { types } from 'util';
 import Secret from '@/components/Secret';
+import Masthead from '@/components/masthead/Masthead';
 
 export default function Home() {
-  const [gradient, setGradient] = useState({ randomColor1: "#fd644f", randomColor2: "#ff9d00"});
+  
   const { data: projects, status: projects_status, error: projects_error } = useQuery(["projects"], async () => await fetch("/api/public/projects").then(async (data) => await data.json().catch(() => [])).catch(() => []));
   const { data: posts, status: posts_status, error: posts_error } = useQuery(["posts"], async () => await fetch("/api/public/posts?limit=3").then(async (data) => await data.json().catch(() => [])).catch((x) => { console.log(x); return [];}));
-  const [typeState, setTypeState] = useState(0);
-  const [mounted, setMounted] = useState(false);
-  useEffect(() => {
-    if (!mounted) setMounted(true)
-    if (mounted && typeState < "Auxdible".length) {
-      setTimeout(() => {
-        setTypeState(typeState + 1);
-      }, 225)
-    }
-  }, [typeState, mounted]);
-  function changeGradient() {
-    setGradient({ randomColor1: '#' + Math.floor(Math.random()*16777215).toString(16), randomColor2: '#' + Math.floor(Math.random()*16777215).toString(16) })
-  }
   return (<>
-  <header className="flex flex-row max-lg:flex-col min-h-screen w-full items-center mx-auto max-lg:my-16 overflow-hidden">
-      <div className={"flex flex-col justify-center gap-5 flex-1 flex-grow font-roboto text-2xl text"}>
-          <section className={"max-lg:text-center mx-auto"}>
-          <h1 className={"text-9xl max-sm:text-6xl pt-4 font-raleway text-primary"}>{"Auxdible".split("").slice(0, typeState)}{mounted && typeState < "Auxdible".length ? <span className={"animate-blink"}>_</span> : ""}</h1>
-          <p className={`text-4xl max-sm:text-3xl font-montserrat py-2 transition-all duration-500 delay-500 ${typeState >= "Auxdible".length ? "opacity-100 translate-x-0" : "opacity-0 -translate-x-32"}`}>Full Stack Developer</p>
-          <p className={`text-2xl max-sm:text-xl font-roboto dark:text-gray-400 text-gray-600 italic transition-all duration-500 delay-1000 ${typeState >= "Auxdible".length ? "opacity-100 translate-x-0" : "opacity-0 translate-x-32"}`}>(a.k.a. Steven Primeaux)</p>
-          </section>
-          
-      </div>
-      <div className={"flex flex-1 flex-grow flex-shrink max-lg:w-screen lg:h-screen overflow-hidden"}>
-        <div className={"w-full"}>
-        { /* I have no idea what this is, how it works, or why it works. I just know it works. 
-        That's all you and I need to know. Sure it makes a js pointerId undefined error that anyone can see in the logs. 
-        but it doesn't affect the site.
-        SO IGNORE IT! */}
-        <Canvas ref={(node) => {
-          if (node) node.style.setProperty("touch-action", "pan-y", "important")
-        }} className={"canvas-wrapper"} style={ { position: "relative"} }>
-          <OrbitControls enableZoom={false} enableDamping={false} enablePan={false} autoRotate autoRotateSpeed={1} />
-          <ambientLight intensity={1} />
-          <perspectiveCamera/>
-          <Suspense fallback={null}>
-            <Glasses frustumCulled={false} rotation={[89,0,0]} {...gradient} />
-          </Suspense>
-        </Canvas>
-        </div>
-      </div>
-    </header>
-    <button onClick={() => changeGradient() } className={"block font-roboto text-3xl text mx-auto w-fit border max-sm:text-xl dark:border-orange-400 hover:scale-105 transition-transform border-orange-700 p-2 rounded-lg mb-72"}>Change glasses color 😎</button>
+    <Masthead/>
     <div className="flex flex-row max-lg:flex-col justify-center items-center align-middle mx-auto mb-40 lg:gap-40 overflow-hidden w-full">
     <motion.h1 initial={{ opacity: 0, transform: "translateX(-8rem)" }}
   whileInView={{ opacity: 1, transform: "translateX(0)" }} transition={{ duration: 1 }} viewport={{ once: true }} className={"text-8xl max-sm:text-5xl pt-4 font-raleway text-primary max-lg:text-center max-lg:my-10 w-fit"}>About<br/>Me</motion.h1>
