+++ components/Navbar.tsx
@@ -11,13 +11,16 @@ export default function Navbar() {
     const { data: session, status } = useSession();
     return (<>
     <div className={`z-10 fixed left-1/2 top-0 -translate-x-1/2 max-w-4xl w-full`}>
-    <nav className={"sticky dark:bg-black bg-gray-100 flex flex-row justify-center max-h-32 align-middle border border-t-0 dark:border-orange-400 border-orange-700 md:rounded-bl-3xl  md:rounded-br-3xl"}>
+    
+    <nav className={"sticky"}>
+        <div className={"absolute -z-10 -inset-[1px] bg-gradient-to-t max-md:bg-gradient-to-r from-orange-400 to-red-500 md:rounded-bl-3xl md:rounded-br-3xl"}></div>
+        <div className={"dark:bg-black bg-gray-100 flex flex-row justify-center max-h-32 align-middle md:rounded-bl-3xl md:rounded-br-3xl"}>
         <div className={"justify-center gap-20 items-center flex flex-1 flex-grow flex-shrink text-xl"}>
             <button aria-label={"Navbar Collapse"} onClick={() => setCollapse(!collapse)} className={"md:hidden max-md:text-xl text-3xl border-2 dark:border-gray-200 border-gray-700 p-2 focus:p-[10px] transition-all rounded-xl"}>
                 <BiMenuAltLeft/>
             </button>
             <span className={"max-md:hidden justify-center gap-20 items-center flex flex-1 flex-grow flex-shrink text-2xl font-montserrat"}>
-            {status == "loading" ? <BiLoaderCircle className={"animate-spin animate-pulse"}/> : session && session.user ? <MiniProfile user={session.user}/> : <Link className={"h-fit hover:text-orange-500 hover:translate-y-1 transition-all"} href={"/auth/signin"}>Sign in</Link>}
+            {status == "loading" ? <BiLoaderCircle className={"animate-spin"}/> : session && session.user ? <MiniProfile user={session.user}/> : <Link className={"h-fit hover:text-orange-500 hover:translate-y-1 transition-all"} href={"/auth/signin"}>Sign in</Link>}
                 
                 <Link className={"h-fit hover:text-orange-500 hover:translate-y-1 transition-all"} href={"/blog"}>Blog</Link>
             </span>
@@ -38,6 +41,8 @@ export default function Navbar() {
             <ThemeButton/>
             
         </div>
+        </div>
+        
     </nav>
     {collapse ? 
     <div className={"sticky -z-10 flex flex-col text-2xl dark:bg-gray-900 bg-gray-200 transition-transform animate-navbarCollapse font-montserrat"}>
