+++ styles/globals.scss
@@ -10,6 +10,9 @@
     .text {
         @apply dark:text-gray-100 text-gray-600
     }
+    .text-title {
+        @apply bg-gradient-to-l from-orange-400 to-red-500 bg-clip-text text-transparent;
+    }
     .text-primary {
         @apply dark:text-orange-400 text-orange-600
     }
@@ -84,9 +87,9 @@
 }
 @keyframes blink {
     from, to {
-        opacity: 0;
+        @apply bg-opacity-0 text-transparent;
     }
     50% {
-        opacity: 1;
+        @apply text-title;
     }
 }
